Introduce Parser context
An opaque value optionally attached to the Parser which can be accessed via the SDT action '$Context'. While the lexer has some notion that a Context might be a 'Sourcer' to obtain a source file from, the parser cares even less and does not use the value for any purpose whatsoever.

Why would parser and lexical context be different?

Outside trivial cases, the lexical context is likely to be more narrowly scoped, while the parser context may be more expansive. When parsing a C-style #include the lexer only cares about the format of the tokens while the parser cares about the their content.

*Alternative approach*

Instead of 'Context', perhaps we could instead make it possible for an adt action to see the parser invoking it, allowing the user to extend the class however they see fit.

```
StructDecl : "struct" identifier ":" identifier "{" StructBody "} << ast.NewStruct($Parser, $1, $3, $5) >>
```